Project - SOCIAL BUZZ Analytics and Visualization
Social Buzz is a leading social media and content creation company which have requested for a comprehensive audit of their big data practices, strategic recommendations for a successful IPO, and an analysis of their most popular content categories. 

Key tasks involve creating a big data best practices presentation, conducting an audit, extracting and analyzing data sets. The goal is to provide Social Buzz with the expertise and guidance needed to manage their scaling challenges and ensure a smooth transition to a public company.